[Tasks] introduce a data pipeline for OSM data gathering
This commit introduces [Luigi](https://luigi.readthedocs.io/en/stable/) for building data pipelines within the current project. The aim is to highlighting task dependencies, and enhancing the data generation process. It increases data generation reliability by keeping a trace of each step on the file system.

This work is still in progress, as the task dependency graph must be clearly defined, and the process must be optimized. For instance, this commit is not equal to legacy code, as there are still NaN column serialization issue with GeoPandas (this case is possible as we serialize transitional states).